Rename and refactor ParameterMap
ParameterMap is now called GenMap for generalised map. The old header
is still available for compatibility and makes GenMap available as
ParameterMap via a typedef.

Please note, that the ParameterMap will soon become deprecated and
eventually removed from the library, so it is a good idea to migrate
your code asap.