Use terminate handlers for ExceptionSystem
Instead of either throwing or aborting the execution we now always
throw and use a custom terminate_handler to obtain the backtrace and
print the exception details. On setup of the system the verbosity can
be configured.

If the system is not properly setup everything still works, but the
exceptions are less detailed.